Update tests and migration toolkit fork
