+++ apps/cqrs-func/utils/__tests__/message.test.ts
@@ -4,9 +4,10 @@ import * as O from "fp-ts/lib/Option";
 import * as TE from "fp-ts/TaskEither";
 import {
   aMessageContent,
-  aRetrievedMessageWithoutContent
+  aRetrievedMessageWithoutContent,
 } from "../../__mocks__/message";
 
+import * as messageUtils from "../message-utils";
 import { enrichMessageContent } from "../message";
 import { vi, beforeEach, describe, test, expect } from "vitest";
 
@@ -15,14 +16,17 @@ const getContentFromBlobMock = vi
   .mockImplementation(() => TE.right(O.some(aMessageContent)));
 const mockMessageModel = {
   find: vi.fn(),
-  getContentFromBlob: getContentFromBlobMock
+  getContentFromBlob: getContentFromBlobMock,
 };
 
-const mockBlobService = {};
+const mockBlobService = {
+  primary: {},
+  secondary: {},
+};
 
 const aRetrievedMessage: RetrievedMessage = {
   ...aRetrievedMessageWithoutContent,
-  kind: "IRetrievedMessageWithoutContent"
+  kind: "IRetrievedMessageWithoutContent",
 };
 
 describe("enrichMessageContent", () => {
@@ -31,10 +35,14 @@ describe("enrichMessageContent", () => {
   });
 
   test("should enrich a message with its' related content", async () => {
+    vi.spyOn(messageUtils, "getContentFromBlob").mockImplementation(() =>
+      TE.right(O.some(aMessageContent)),
+    );
+
     const result = await enrichMessageContent(
       mockMessageModel as any,
       mockBlobService as any,
-      aRetrievedMessage
+      aRetrievedMessage,
     )();
 
     expect(E.isRight(result)).toBeTruthy();
@@ -42,45 +50,49 @@ describe("enrichMessageContent", () => {
       expect(result.right).toEqual({
         ...aRetrievedMessage,
         content: aMessageContent,
-        kind: "IRetrievedMessageWithContent"
+        kind: "IRetrievedMessageWithContent",
       });
     }
   });
 
   test("should return a retriable error if blob storage cannot retrieve the message content", async () => {
-    getContentFromBlobMock.mockImplementationOnce(() =>
-      TE.left(new Error("cannot reach blob storage"))
+    vi.spyOn(messageUtils, "getContentFromBlob").mockImplementation(() =>
+      TE.left(new Error("cannot reach blob storage")),
     );
+
     const result = await enrichMessageContent(
       mockMessageModel as any,
       mockBlobService as any,
-      aRetrievedMessage
+      aRetrievedMessage,
     )();
 
     expect(E.isLeft(result)).toBeTruthy();
     if (E.isLeft(result)) {
       expect(result.left).toEqual(
         expect.objectContaining({
-          retriable: true
-        })
+          retriable: true,
+        }),
       );
     }
   });
 
   test("should return a not retriable error if message content cannot be found", async () => {
-    getContentFromBlobMock.mockImplementationOnce(() => TE.right(O.none));
+    vi.spyOn(messageUtils, "getContentFromBlob").mockImplementation(() =>
+      TE.right(O.none),
+    );
+
     const result = await enrichMessageContent(
       mockMessageModel as any,
       mockBlobService as any,
-      aRetrievedMessage
+      aRetrievedMessage,
     )();
 
     expect(E.isLeft(result)).toBeTruthy();
     if (E.isLeft(result)) {
       expect(result.left).toEqual(
         expect.objectContaining({
-          retriable: false
-        })
+          retriable: false,
+        }),
       );
     }
   });
